Use a systemd-enabled Debian 9 Docker image
Since the particular image we are using is already Ansible-ready, we
can mark it as "pre_build_image: yes" and save a little time.

Note also that we have to give the container some extra capabilities
and mount (read-only) /sys/fs/cgroup for systemd to work inside of the
Docker container.

See these links for more details:
* https://molecule.readthedocs.io/en/stable/examples.html#systemd-container
* https://www.jeffgeerling.com/blog/2018/testing-your-ansible-roles-molecule